+++ src/main/java/subway/dto/request/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 import javax.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
