+++ src/main/java/subway/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 import java.util.HashMap;
 import java.util.Map;
