+++ src/main/java/subway/exception/ExceptionAdvice.java
@@ -6,7 +6,7 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import subway.dto.ErrorResponse;
+import subway.dto.response.ErrorResponse;
 
 @RestControllerAdvice
 @Slf4j
