+++ src/main/java/subway/ui/SectionController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.SectionService;
-import subway.dto.SectionRequest;
+import subway.dto.request.SectionRequest;
 import subway.dto.StationDto;
 import subway.dto.StationsDto;
