+++ src/test/java/subway/integration/SectionIntegrationTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.LineRequest;
+import subway.dto.request.LineRequest;
 
 @DisplayName("지하철역 관련 기능")
 public class SectionIntegrationTest extends IntegrationTest {
